Выполнены задания к Уроку 2
