+++ 3-STUDENTS/georgedanilyants/project/public/index.html
@@ -6,7 +6,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Clothes Store</title>
     <link href="https://fonts.googleapis.com/css2?family=Lato:wght@100;300;400;700;900&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="../src/layout/styles/style.css">
     <script src="https://kit.fontawesome.com/8d73d6a795.js" crossorigin="anonymous"></script>
 </head>
 
@@ -345,11 +344,6 @@
         <!-- Сюда рендерятся картинки -->
     </div>
 </div>
-    <script src="../src/components/app.js"></script>
-    <script src="../src/components/cart.js"></script>
-    <script src="../src/components/catalog.js"></script>
-    <script>catalog.init();</script>
-    <script src="../src/components/gallery.js"></script>
 </body>
 
 </html>
