+++ 3-STUDENTS/georgedanilyants/project/src/components/index.js
@@ -1,7 +1,7 @@
-import basket from './cart.js' 
-import catalog from './catalog.js'
+import Basket from './basket.js' 
+import Catalog from './catalog.js'
 
 export default () => {
-    basket.init();
-    catalog.init();
+   let basket = new Basket();
+   let catalog = new Catalog(basket);
 }
