+++ 3-STUDENTS/georgedanilyants/project/src/index.js
@@ -1,3 +1,6 @@
 import app from './components'
+import 'bootstrap';
+import 'bootstrap/dist/css/bootstrap.min.css'
+import './layout/styles/style.css'
 
 app();
